+++ app/src/main/res/layout/fragment_details_kitchen.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.DetailsKitchenFragment">
+    tools:context=".ui.detailsKitchen.DetailsKitchenFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/kitchen_toolbar"
